+++ lib/ui/widget/svg/svgs.dart
@@ -1592,12 +1592,6 @@ class SvgIcons {
     height: 19,
   );
 
-  static const SvgData logo = SvgData(
-    'assets/icons/logo.svg',
-    width: 21.78,
-    height: 25,
-  );
-
   static const SvgData workRust = SvgData(
     'assets/icons/work_rust.svg',
     width: 32,
@@ -2526,5 +2520,23 @@ class SvgIcons {
     'assets/icons/arrow_up.svg',
     width: 12,
     height: 7,
+  );
+
+  static const SvgData shareBig = SvgData(
+    'assets/icons/share_big.svg',
+    width: 23,
+    height: 29,
+  );
+
+  static const SvgData addToDock = SvgData(
+    'assets/icons/add_to_dock.svg',
+    width: 21,
+    height: 17,
+  );
+
+  static const SvgData addToHomeScreen = SvgData(
+    'assets/icons/add_to_home_screen.svg',
+    width: 17,
+    height: 17,
   );
 }